requirements/base.txt
Django==2.1
Pillow==5.2.0
awesome-slugify==1.6.5
celery==4.2.1
git+https://github.com/justinmayer/django-autoslug.git@master
django-braces==1.13.0
django-ckeditor==5.6.1
django-countries==5.3.1
django-grappelli==2.11.1
django-imagekit==4.0.2
django-redis==4.9.0
djangorestframework==3.8.2
gunicorn==19.9.0
html5lib==1.0.1
psycopg2-binary==2.7.5
redis==2.10.6
beautifulsoup4==4.6.1
mutagen==1.41.0
requests==2.19.1
django-cleanup==2.1.0
textile==3.0.3
social-auth-app-django==2.1.0
social-auth-core==1.7.0
raven==6.9.0
django-environ==0.4.5
